# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/tokens": {
            "get": {
                "description": "Fetch the token details by address with optional platform param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Get token details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key of the client",
                        "name": "api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Platform",
                        "name": "platform",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TokenDetails"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.RestError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.RestError"
                        }
                    }
                }
            }
        },
        "/v1/tokens/multi": {
            "get": {
                "description": "Fetch multiple token details with optional platform param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Multi get token details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key of the client",
                        "name": "api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma separated addresses",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Platform",
                        "name": "platform",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MultiTokenDetails"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.RestError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.RestError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.MultiTokenDetails": {
            "type": "object",
            "properties": {
                "found_count": {
                    "type": "integer"
                },
                "not_found_count": {
                    "type": "integer"
                },
                "results": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/dto.TokenDetails"
                    }
                }
            }
        },
        "dto.RestError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {},
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "dto.TokenDetails": {
            "type": "object",
            "properties": {
                "logo": {
                    "$ref": "#/definitions/models.Logo"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "token_platform": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TokenPlatform"
                    }
                },
                "usd_24_hour_change": {
                    "type": "number"
                },
                "usd_24_hour_volume": {
                    "type": "number"
                },
                "usd_market_cap": {
                    "type": "number"
                },
                "usd_price": {
                    "type": "number"
                }
            }
        },
        "dto.TokenPlatform": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "decimal": {
                    "type": "integer"
                },
                "platform_name": {
                    "type": "string"
                }
            }
        },
        "models.Logo": {
            "type": "object",
            "properties": {
                "large": {
                    "type": "string"
                },
                "small": {
                    "type": "string"
                },
                "thumb": {
                    "type": "string"
                }
            }
        }
    }
}